+++ Toolkit/RSPiX/ORANGE/GUI/ProcessGui.cpp
@@ -82,7 +82,7 @@
 //////////////////////////////////////////////////////////////////////////////
 
 // ID of last item pressed.
-long RProcessGui::ms_lPressedId;
+int32_t RProcessGui::ms_lPressedId;
 
 //////////////////////////////////////////////////////////////////////////////
 // Helper Functions.
@@ -254,7 +254,7 @@ void RProcessGui::SetGuisToNotify(	// Returns nothing.
 // with an ID other than 0 or the update callback, m_fnUpdate, if any,
 // returns non-zero.
 //////////////////////////////////////////////////////////////////////////////
-long RProcessGui::DoModal(				// Returns ID of pressed GUI that terminated 
+int32_t RProcessGui::DoModal(				// Returns ID of pressed GUI that terminated
 												// modal loop or value returned from 
 												// m_fnUpdate, if any.
 	RGuiItem* pgui,						// In:  GUI to be processed or nullptr.
@@ -309,7 +309,7 @@ long RProcessGui::DoModal(				// Returns ID of pressed GUI that terminated
 // Call this to process a GUI modelessly.  This function processes the
 // GUI for only one iteration allowing the caller continuous control.
 //////////////////////////////////////////////////////////////////////////////
-long RProcessGui::DoModeless(		// Returns ID of pressed GUI or value.
+int32_t RProcessGui::DoModeless(		// Returns ID of pressed GUI or value.
 	RGuiItem* pgui,					// In:  GUI to be processed or nullptr.
 	RInputEvent* pie,					// In:  Input event to process.
 	RImage* pimDst /*= nullptr */)		// Where to draw dialog and rspBlit from.
